+++ test/test_sync_subscriptions.cpp
@@ -156,13 +156,19 @@ TEST(Sync_SubscriptionStoreStateUpdates)
     {
         auto set = store.get_latest().make_mutable_copy();
         CHECK_EQUAL(set.size(), 1);
-        set.insert_or_assign(query_a);
+        // This is just to create a unique name for this sub so we can verify that the iterator returned by
+        // insert_or_assign is pointing to the subscription that was just created.
+        std::string new_sub_name = ObjectId::gen().to_string();
+        auto&& [inserted_it, inserted] = set.insert_or_assign(new_sub_name, query_a);
+        CHECK(inserted);
+        CHECK_EQUAL(inserted_it->name(), new_sub_name);
         CHECK_EQUAL(set.size(), 2);
         auto it = set.begin();
+        CHECK_EQUAL(it->name(), "b sub");
         it = set.erase(it);
         CHECK_NOT_EQUAL(it, set.end());
         CHECK_EQUAL(set.size(), 1);
-        CHECK_EQUAL(it->name(), "b sub");
+        CHECK_EQUAL(it->name(), new_sub_name);
         it = set.erase(it);
         CHECK_EQUAL(it, set.end());
         CHECK_EQUAL(set.size(), 0);