+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * Renamed ClientResyncMode::SeamlessLoss -> DiscardLocal.
 * Updated sync client to be able to open connections to FLX sync-enabled apps in baas ([#5009](https://github.com/realm/realm-core/pull/5009))
 * SubscriptionSet::erase now returns the correct itererator for the "next" subscription ([#5053](https://github.com/realm/realm-core/pull/5053))
+* SubscriptionSet::insert_or_assign now returns an iterator pointing to the correct subscription ([#5049](https://github.com/realm/realm-core/pull/5049))
 
 ----------------------------------------------